requirements.txt
Pillow
beautifulsoup4
argparse
